+++ src/components/api/const/interfaces/index.tsx
@@ -11,5 +11,7 @@ import { FilterValues } from "./facilityFilterInterface";
 import { UpdateProfileInterface, UpdatedProfileData } from "./profileUpdateInterface";
 import { NavBarProps } from "./navBarInterfaces";
 import { ContactFormInputs } from "./contactFormInterface";
+import { SearchFormInputFP } from "./searchInterface";
+import { SearchFormMainProps } from "./convertedSearchInterface";
 
-export type { SingleVenueResponse, SingleVenueResponseContent, SingleVenueProp, BookingCalendarProps, bookingDate, CreateVenueFormInputs, VenueInterface, bookingTileInterface, SearchFormInputInterface, ConvertedSearchDataInterface, SearchReturnInterface, Bookings, FilterValues, UpdateProfileInterface, UpdatedProfileData, NavBarProps, ContactFormInputs };
+export type { SingleVenueResponse, SingleVenueResponseContent, SingleVenueProp, BookingCalendarProps, bookingDate, CreateVenueFormInputs, VenueInterface, bookingTileInterface, SearchFormInputInterface, ConvertedSearchDataInterface, SearchReturnInterface, Bookings, FilterValues, UpdateProfileInterface, UpdatedProfileData, NavBarProps, ContactFormInputs, SearchFormInputFP, SearchFormMainProps };
