+++ src/components/api/const/interfaces/searchInterface.tsx
@@ -1,0 +1,6 @@
+export interface SearchFormInputFP {
+  location: string;
+  dateFrom: string;
+  dateTo: string;
+  guests: number;
+}
